% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfte.R
\name{ready_fonts}
\alias{ready_fonts}
\title{Ready fonts for bfte theme options}
\usage{
ready_fonts(
  google_font_name_vector = c("Roboto", "Nothing You Could Do", "Major Mono Display"),
  font_ref_vector = c("roboto", "nothing", "majormono")
)
}
\arguments{
\item{google_font_name_vector}{vector of Google font names}

\item{font_ref_vector}{vector of abbreviated font names}
}
\value{
fonts loaded into namespace
}
\description{
Loads google fonts via the `showtext`` package.
}
\details{
Three fonts are loaded by default: *Roboto*, *Nothing You Could Do*, and *Major Mono Display*.
}
\examples{
ready_fonts()
}
\seealso{
\code{\link{theme_bfte_roboto}}, \code{\link{theme_bfte_nothing}}, \code{\link{theme_bfte_majormono}}
}
