% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfte.R
\name{theme_bfte_dx_majormono}
\alias{theme_bfte_dx_majormono}
\title{BFTE theme with *Major Mono Display* font}
\usage{
theme_bfte_dx_majormono()
}
\value{
theme added to a `ggplot2` object
}
\description{
Adds a themed appearance and guide to a `ggplot2` object with a **discrete** x-axis variable.
}
\details{
The *Major Mono Display* font is used in the theme.
}
\examples{
ggplot2::ggplot(data = data.frame(x = c("zero","ten","seventeen"), y = c(79,11,18)),
 ggplot2::aes(x = x, y = y)
 ) +
 ggplot2::geom_point() +
 theme_bfte_dx_majormono()
}
\seealso{
\code{\link{ready_fonts}}, \code{\link{theme_bfte_dx_roboto}}, \code{\link{theme_bfte_dx_nothing}}
}
