% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfte.R
\docType{data}
\name{theme_bfte_roboto}
\alias{theme_bfte_roboto}
\title{BFTE theme with Roboto font}
\format{
An object of class \code{list} of length 2.
}
\usage{
theme_bfte_roboto
}
\value{
added to a
}
\description{
Adds a themed appearance and guide to a `ggplot2` object.
}
\details{
The Roboto font is used in the theme.
}
\examples{
ggplot(data = data(mtcars), aes(x = cyl, y = mpg)) + geom_point() + theme_bfte_roboto()
}
\seealso{
\code{\link{ready_fonts}}, \code{\link{theme_bfte_nothing}}, \code{\link{theme_bfte_majormono}}
}
\keyword{datasets}
